requirement-analytics.txt
networkx==3.2.1
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
kafka-python==2.0.2
